Refactor bloemenwinkel component structure and enhance styling for improved user experience
